+++ CleverTapSDK/InApps/CTInAppHTMLViewController.m
@@ -230,34 +230,18 @@ typedef enum {
         return;
     }
     
-    NSMutableDictionary *mutableParams = [NSMutableDictionary new];
     NSString *urlString = [navigationAction.request.URL absoluteString];
     NSURL *dl = [NSURL URLWithString:urlString];
+    NSMutableDictionary *mutableParams = [CTInAppUtils getParametersFromURL:urlString];
     
-    // Try to extract the parameters from the URL and overrite default dl if applicable
-    NSMutableDictionary *params = [[NSMutableDictionary alloc] init];
-    NSArray *comps = [urlString componentsSeparatedByString:@"?"];
-    if ([comps count] >= 2) {
-        NSString *query = comps[1];
-        for (NSString *param in [query componentsSeparatedByString:@"&"]) {
-            NSArray *elts = [param componentsSeparatedByString:@"="];
-            if ([elts count] < 2) continue;
-            params[elts[0]] = [elts[1] stringByRemovingPercentEncoding];
-        };
-        mutableParams = [params mutableCopy];
-        NSString *c2a = params[CLTAP_PROP_WZRK_CTA];
-        if (c2a) {
-            c2a = [c2a stringByRemovingPercentEncoding];
-            NSArray *parts = [c2a componentsSeparatedByString:@"__dl__"];
-            if (parts && [parts count] == 2) {
-                dl = [NSURL URLWithString:parts[1]];
-                mutableParams[CLTAP_PROP_WZRK_CTA] = parts[0];
-            }
-        }
+    // Use the url from the callToAction param to update action
+    if (mutableParams[@"deeplink"]) {
+        dl = mutableParams[@"deeplink"];
     }
+    
     if (self.delegate && [self.delegate respondsToSelector:@selector(handleNotificationAction:forNotification:withExtras:)]) {
         CTNotificationAction *action = [[CTNotificationAction alloc] initWithOpenURL:dl];
-        [self.delegate handleNotificationAction:action forNotification:self.notification withExtras:mutableParams];
+        [self.delegate handleNotificationAction:action forNotification:self.notification withExtras:mutableParams[@"params"]];
     }
     [self hide:YES];
     decisionHandler(WKNavigationActionPolicyCancel);
@@ -510,29 +494,6 @@ typedef enum {
     }
 }
 
-- (void)hideFromWindow:(BOOL)animated {
-    void (^completionBlock)(void) = ^ {
-        [self->webView.configuration.userContentController removeScriptMessageHandlerForName:@"clevertap"];
-        [self.window removeFromSuperview];
-        self.window = nil;
-        if (self.delegate && [self.delegate respondsToSelector:@selector(notificationDidDismiss:fromViewController:)]) {
-            [self.delegate notificationDidDismiss:self.notification fromViewController:self];
-        }
-    };
-    
-    if (animated) {
-        [UIView animateWithDuration:0.25 animations:^{
-            self.window.alpha = 0;
-        } completion:^(BOOL finished) {
-            completionBlock();
-        }];
-    }
-    else {
-        completionBlock();
-    }
-}
-
-
 #pragma mark - Public
 
 - (void)show:(BOOL)animated {
@@ -541,7 +502,14 @@ typedef enum {
 }
 
 - (void)hide:(BOOL)animated {
-    [self hideFromWindow:animated];
+    __weak typeof(self) weakSelf = self;
+    [self hideFromWindow:animated withCompletion:^{
+        __strong typeof(weakSelf) strongSelf = weakSelf;
+        if (!strongSelf) {
+            return;
+        }
+        [strongSelf->webView.configuration.userContentController removeScriptMessageHandlerForName:@"clevertap"];
+    }];
 }
 
 @end